+++ drivers/storage/portworx/deployment_test.go
@@ -961,7 +961,7 @@ func TestPodSpecWithCloudStorageSpec(t *testing.T) {
 	)
 	nodeName := "testNode"
 
-	zoneToInstancesMap := map[string]int{"a": 3, "b": 3, "c": 2}
+	zoneToInstancesMap := map[string]uint64{"a": 3, "b": 3, "c": 2}
 	driver := portworx{
 		k8sClient:          k8sClient,
 		recorder:           record.NewFakeRecorder(0),
@@ -1122,13 +1122,13 @@ func TestPodSpecWithCloudStorageSpec(t *testing.T) {
 	cluster.Spec.CloudStorage = &corev1.CloudStorageSpec{
 		CapacitySpecs: []corev1.CloudStorageCapacitySpec{
 			{
-				MinIOPS:          uint32(100),
+				MinIOPS:          uint64(100),
 				MinCapacityInGiB: uint64(100),
 				MaxCapacityInGiB: uint64(200),
 				Options:          map[string]string{"foo1": "bar1"},
 			},
 			{
-				MinIOPS:          uint32(200),
+				MinIOPS:          uint64(200),
 				MinCapacityInGiB: uint64(200),
 				MaxCapacityInGiB: uint64(500),
 				Options:          map[string]string{"foo3": "bar3"},
@@ -1140,13 +1140,13 @@ func TestPodSpecWithCloudStorageSpec(t *testing.T) {
 		{
 			Type:      "foo",
 			SizeInGiB: uint64(120),
-			IOPS:      uint32(110),
+			IOPS:      uint64(110),
 			Options:   map[string]string{"foo1": "bar1"},
 		},
 		{
 			Type:      "bar",
 			SizeInGiB: uint64(220),
-			IOPS:      uint32(210),
+			IOPS:      uint64(210),
 			Options: map[string]string{
 				"foo3": "bar3",
 			}},
@@ -1160,19 +1160,19 @@ func TestPodSpecWithCloudStorageSpec(t *testing.T) {
 		"-max_storage_nodes_per_zone", "2",
 	}
 
-	inputInstancesPerZone := 2
+	inputInstancesPerZone := uint64(2)
 	mockStorageManager.EXPECT().
 		GetStorageDistribution(&cloudops.StorageDistributionRequest{
-			ZoneCount:        len(zoneToInstancesMap),
+			ZoneCount:        uint64(len(zoneToInstancesMap)),
 			InstancesPerZone: inputInstancesPerZone,
 			UserStorageSpec: []*cloudops.StorageSpec{
 				{
-					IOPS:        uint32(100),
+					IOPS:        uint64(100),
 					MinCapacity: uint64(100),
 					MaxCapacity: uint64(200),
 				},
 				{
-					IOPS:        uint32(200),
+					IOPS:        uint64(200),
 					MinCapacity: uint64(200),
 					MaxCapacity: uint64(500),
 				},
@@ -1185,14 +1185,14 @@ func TestPodSpecWithCloudStorageSpec(t *testing.T) {
 					DriveType:        "foo",
 					DriveCount:       1,
 					InstancesPerZone: inputInstancesPerZone,
-					IOPS:             uint32(110),
+					IOPS:             uint64(110),
 				},
 				{
 					DriveCapacityGiB: uint64(220),
 					DriveType:        "bar",
 					DriveCount:       1,
 					InstancesPerZone: inputInstancesPerZone,
-					IOPS:             uint32(210),
+					IOPS:             uint64(210),
 				},
 			},
 		}, nil)
@@ -1213,13 +1213,13 @@ func TestPodSpecWithCloudStorageSpec(t *testing.T) {
 	cluster.Spec.CloudStorage = &corev1.CloudStorageSpec{
 		CapacitySpecs: []corev1.CloudStorageCapacitySpec{
 			{
-				MinIOPS:          uint32(100),
+				MinIOPS:          uint64(100),
 				MinCapacityInGiB: uint64(100),
 				MaxCapacityInGiB: uint64(200),
 				Options:          map[string]string{"foo1": "bar1"},
 			},
 			{
-				MinIOPS:          uint32(200),
+				MinIOPS:          uint64(200),
 				MinCapacityInGiB: uint64(200),
 				MaxCapacityInGiB: uint64(500),
 				Options:          map[string]string{"foo3": "bar3"},
@@ -1380,13 +1380,13 @@ func TestPodSpecWithCapacitySpecsAndDeviceSpecs(t *testing.T) {
 		},
 		CapacitySpecs: []corev1.CloudStorageCapacitySpec{
 			{
-				MinIOPS:          uint32(100),
+				MinIOPS:          uint64(100),
 				MinCapacityInGiB: uint64(100),
 				MaxCapacityInGiB: uint64(200),
 				Options:          map[string]string{"foo1": "bar1"},
 			},
 			{
-				MinIOPS:          uint32(200),
+				MinIOPS:          uint64(200),
 				MinCapacityInGiB: uint64(200),
 				MaxCapacityInGiB: uint64(500),
 				Options:          map[string]string{"foo3": "bar3"},
@@ -1394,7 +1394,7 @@ func TestPodSpecWithCapacitySpecsAndDeviceSpecs(t *testing.T) {
 		},
 	}
 
-	zoneToInstancesMap := map[string]int{"a": 3, "b": 3, "c": 2}
+	zoneToInstancesMap := map[string]uint64{"a": 3, "b": 3, "c": 2}
 	driver := portworx{
 		k8sClient:          k8sClient,
 		recorder:           record.NewFakeRecorder(0),
@@ -1402,19 +1402,19 @@ func TestPodSpecWithCapacitySpecsAndDeviceSpecs(t *testing.T) {
 		cloudProvider:      "mock",
 	}
 
-	inputInstancesPerZone := 2
+	inputInstancesPerZone := uint64(2)
 	mockStorageManager.EXPECT().
 		GetStorageDistribution(&cloudops.StorageDistributionRequest{
-			ZoneCount:        len(zoneToInstancesMap),
+			ZoneCount:        uint64(len(zoneToInstancesMap)),
 			InstancesPerZone: inputInstancesPerZone,
 			UserStorageSpec: []*cloudops.StorageSpec{
 				{
-					IOPS:        uint32(100),
+					IOPS:        uint64(100),
 					MinCapacity: uint64(100),
 					MaxCapacity: uint64(200),
 				},
 				{
-					IOPS:        uint32(200),
+					IOPS:        uint64(200),
 					MinCapacity: uint64(200),
 					MaxCapacity: uint64(500),
 				},
@@ -1427,14 +1427,14 @@ func TestPodSpecWithCapacitySpecsAndDeviceSpecs(t *testing.T) {
 					DriveType:        "foo",
 					DriveCount:       1,
 					InstancesPerZone: inputInstancesPerZone,
-					IOPS:             uint32(110),
+					IOPS:             uint64(110),
 				},
 				{
 					DriveCapacityGiB: uint64(220),
 					DriveType:        "bar",
 					DriveCount:       1,
 					InstancesPerZone: inputInstancesPerZone,
-					IOPS:             uint32(210),
+					IOPS:             uint64(210),
 				},
 			},
 		}, nil)
@@ -2924,7 +2924,7 @@ func TestStorageNodeConfig(t *testing.T) {
 	nodeName2 := "testNode2"
 	nodeName3 := "testNode3"
 
-	zoneToInstancesMap := map[string]int{"a": 3, "b": 3, "c": 2}
+	zoneToInstancesMap := map[string]uint64{"a": 3, "b": 3, "c": 2}
 	driver := portworx{
 		k8sClient:          k8sClient,
 		recorder:           record.NewFakeRecorder(0),
@@ -2954,13 +2954,13 @@ func TestStorageNodeConfig(t *testing.T) {
 	cluster.Spec.CloudStorage = &corev1.CloudStorageSpec{
 		CapacitySpecs: []corev1.CloudStorageCapacitySpec{
 			{
-				MinIOPS:          uint32(100),
+				MinIOPS:          uint64(100),
 				MinCapacityInGiB: uint64(100),
 				MaxCapacityInGiB: uint64(200),
 				Options:          map[string]string{"foo1": "bar1"},
 			},
 			{
-				MinIOPS:          uint32(200),
+				MinIOPS:          uint64(200),
 				MinCapacityInGiB: uint64(200),
 				MaxCapacityInGiB: uint64(500),
 				Options:          map[string]string{"foo3": "bar3"},
@@ -2976,19 +2976,19 @@ func TestStorageNodeConfig(t *testing.T) {
 		"-max_storage_nodes_per_zone", "2",
 	}
 
-	inputInstancesPerZone := 2
+	inputInstancesPerZone := uint64(2)
 	mockStorageManager.EXPECT().
 		GetStorageDistribution(&cloudops.StorageDistributionRequest{
-			ZoneCount:        len(zoneToInstancesMap),
+			ZoneCount:        uint64(len(zoneToInstancesMap)),
 			InstancesPerZone: inputInstancesPerZone,
 			UserStorageSpec: []*cloudops.StorageSpec{
 				{
-					IOPS:        uint32(100),
+					IOPS:        uint64(100),
 					MinCapacity: uint64(100),
 					MaxCapacity: uint64(200),
 				},
 				{
-					IOPS:        uint32(200),
+					IOPS:        uint64(200),
 					MinCapacity: uint64(200),
 					MaxCapacity: uint64(500),
 				},
@@ -3001,14 +3001,14 @@ func TestStorageNodeConfig(t *testing.T) {
 					DriveType:        "foo",
 					DriveCount:       1,
 					InstancesPerZone: inputInstancesPerZone,
-					IOPS:             uint32(110),
+					IOPS:             uint64(110),
 				},
 				{
 					DriveCapacityGiB: uint64(220),
 					DriveType:        "bar",
 					DriveCount:       1,
 					InstancesPerZone: inputInstancesPerZone,
-					IOPS:             uint32(210),
+					IOPS:             uint64(210),
 				},
 			},
 		}, nil)
@@ -3016,7 +3016,7 @@ func TestStorageNodeConfig(t *testing.T) {
 	actual, err = driver.GetStoragePodSpec(cluster, nodeName)
 	assert.NoError(t, err, "Unexpected error on GetStoragePodSpec")
 	assert.ElementsMatch(t, expectedArgs, actual.Containers[0].Args)
-	assert.Equal(t, int32(inputInstancesPerZone), cluster.Status.Storage.StorageNodesPerZone, "wrong number of instances per zone in a the cluster status")
+	assert.Equal(t, uint64(inputInstancesPerZone), cluster.Status.Storage.StorageNodesPerZone, "wrong number of instances per zone in a the cluster status")
 
 	list, err = driver.storageNodesList(cluster)
 	assert.NoError(t, err, "Unexpected error on GetStorageNodeList")
@@ -3026,13 +3026,13 @@ func TestStorageNodeConfig(t *testing.T) {
 		{
 			Type:      "foo",
 			SizeInGiB: uint64(120),
-			IOPS:      uint32(110),
+			IOPS:      uint64(110),
 			Options:   map[string]string{"foo1": "bar1"},
 		},
 		{
 			Type:      "bar",
 			SizeInGiB: uint64(220),
-			IOPS:      uint32(210),
+			IOPS:      uint64(210),
 			Options: map[string]string{
 				"foo3": "bar3",
 			}},
@@ -3051,12 +3051,12 @@ func TestStorageNodeConfig(t *testing.T) {
 	actual, err = driver.GetStoragePodSpec(cluster, nodeName2)
 	assert.NoError(t, err, "Unexpected error on GetStoragePodSpec")
 	assert.ElementsMatch(t, expectedArgs, actual.Containers[0].Args)
-	assert.Equal(t, int32(inputInstancesPerZone), cluster.Status.Storage.StorageNodesPerZone, "wrong number of instances per zone in a the cluster status")
+	assert.Equal(t, uint64(inputInstancesPerZone), cluster.Status.Storage.StorageNodesPerZone, "wrong number of instances per zone in a the cluster status")
 
 	list, err = driver.storageNodesList(cluster)
 	assert.NoError(t, err, "Unexpected error on GetStorageNodeList")
 	assert.Equal(t, 2, len(list), "expected storage nodes in list")
-	assert.Equal(t, int32(inputInstancesPerZone), cluster.Status.Storage.StorageNodesPerZone, "wrong number of instances per zone in a the cluster status")
+	assert.Equal(t, uint64(inputInstancesPerZone), cluster.Status.Storage.StorageNodesPerZone, "wrong number of instances per zone in a the cluster status")
 
 	for _, node := range list {
 		assert.ElementsMatch(t, node.Spec.CloudStorage.DriveConfigs, expectedCloudStorageSpec)
@@ -3067,7 +3067,7 @@ func TestStorageNodeConfig(t *testing.T) {
 	list, err = driver.storageNodesList(cluster)
 	assert.NoError(t, err, "Unexpected error on GetStorageNodeList")
 	assert.Equal(t, 3, len(list), "expected storage nodes in list")
-	assert.Equal(t, int32(inputInstancesPerZone), cluster.Status.Storage.StorageNodesPerZone, "wrong number of instances per zone in a the cluster status")
+	assert.Equal(t, uint64(inputInstancesPerZone), cluster.Status.Storage.StorageNodesPerZone, "wrong number of instances per zone in a the cluster status")
 
 	for _, node := range list {
 		assert.ElementsMatch(t, node.Spec.CloudStorage.DriveConfigs, expectedCloudStorageSpec)
